chore: deduplicate IFlags
